Handle profile fetching after login with profile-loader
