+++ packages/nav-sidebar/middleware.test.js
@@ -1,4 +1,4 @@
-import { actionTypes, actions } from '@bufferapp/async-data-fetch';
+import { actionTypes } from '@bufferapp/async-data-fetch';
 import {
   actionTypes as profileActionTypes,
   actions as profileActions,
@@ -47,18 +47,6 @@ describe('middleware', () => {
       type: 'TEST',
     };
     invoke(action);
-    expect(next).toHaveBeenCalledWith(action);
-  });
-
-  it('should dispatch the data fetch action for fetching the profiles once login has been successful', () => {
-    const { store, next, invoke } = getMiddlewareElements();
-    const action = {
-      type: `login_${actionTypes.FETCH_SUCCESS}`,
-    };
-    invoke(action);
-    expect(store.dispatch).toHaveBeenCalledWith(actions.fetch({
-      name: 'profiles',
-    }));
     expect(next).toHaveBeenCalledWith(action);
   });
