+++ packages/profile-loader/index.js
@@ -10,3 +10,5 @@ export default connect(
 
 // export reducer, actions and action types
 export reducer, { actions, actionTypes } from './reducer';
+
+export middleware from './middleware';
